+++ samples/NHotPhrase.Wpf.Demo/MainWindow.xaml
@@ -20,7 +20,7 @@
                   Unchecked="CheckBoxUnchecked"
                   Content="Global hotkeys enabled" 
                   />
-        <TextBlock Grid.Row="1" HorizontalAlignment="Center" VerticalAlignment="Center" FontSize="48" Text="{Binding Value}" />
+        <TextBlock Grid.Row="1" HorizontalAlignment="Center" VerticalAlignment="Center" FontSize="48" Name="CurrentValue" Text="0" />
         <TextBlock Grid.Row="2" HorizontalAlignment="Left" Text="Toggle Global HotPhrases" />
         <TextBlock Grid.Row="3" HorizontalAlignment="Left" Text="Increment" />
         <TextBlock Grid.Row="4" HorizontalAlignment="Left" Text="Decrement" />
