+++ src/NHotPhrase.WindowsForms/HotPhraseManagerForWinForms.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Windows.Forms;
@@ -54,5 +55,29 @@ namespace NHotPhrase.WindowsForms
                 ? keyword.SendKeysText()
                 : pKey.ToString();
         }
+
+        public List<string> MakeReadyForSending(string target, int splitLength = 8)
+        {
+            if (string.IsNullOrEmpty(target))
+                return new List<string>();
+
+            foreach (var keyword in KeySender.SendKeyEntries.Where(k => !string.IsNullOrEmpty(k.ReplaceWith)))
+                target = target.Replace(keyword.Name, "⌂" + keyword.ReplaceWith + "⌂");
+
+            var list = target.Split('⌂', StringSplitOptions.RemoveEmptyEntries).ToList();
+            while (list.Any(p => p.Length > splitLength))
+                for (var i = 0; i < list.Count; i++)
+                {
+                    if (list[i].Length <= splitLength) continue;
+
+                    var pieces = list[i].SplitInTwo();
+                    list.RemoveAt(i);
+                    list.InsertRange(i, pieces);
+                }
+
+            return list;
+        }
+
+
     }
 }
